+++ sshx/__init__.py
@@ -1,5 +1,5 @@
 
-from __future__ import unicode_literals
+from __future__ import nested_scopes, generators, division, absolute_import, with_statement, print_function, unicode_literals
 
 __version__ = '0.1.1'