+++ sshx/const.py
@@ -1,3 +1,5 @@
+from __future__ import nested_scopes, generators, division, absolute_import, with_statement, print_function, unicode_literals
+
 DEFAULT_USER = 'root'
 DEFAULT_HOST = None
 DEFAULT_PORT = '22'